% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_num.R
\name{big_num}
\alias{big_num}
\title{Infinite precision natural number using a singly linked list}
\usage{
big_num(num = "")
}
\arguments{
\item{num}{A string representation of a num.}
}
\description{
TODO: WRITE THIS

The \code{big_num} class is essentially a wrapper around \link{linked_list}
with a custom print method as well as some defined operators such as
\code{+}, \code{*}, and \code{^} with an integer.
}
