% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigfloat.R
\name{bigfloat}
\alias{bigfloat}
\alias{as_bigfloat}
\alias{is_bigfloat}
\alias{format.bignum_bigfloat}
\title{High-Precision Numeric Vectors}
\usage{
bigfloat(x = character())

as_bigfloat(x, ...)

is_bigfloat(x)

\method{format}{bignum_bigfloat}(x, scientific = FALSE, ...)
}
\arguments{
\item{x}{Object to be coerced or tested}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Creates or tests for high-precision floating-point numeric vectors.
}
\examples{
bigfloat(1 / 3)
bigfloat(1) / 3
}
\seealso{
\code{\link[=biginteger]{biginteger()}}, \code{\link[=double]{double()}}
}
