% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compare.R
\name{bignum-compare}
\alias{bignum-compare}
\title{Comparison operations}
\value{
A logical vector.
}
\description{
\link{biginteger} and \link{bigfloat} vectors support the standard comparison operations.
The base R documentation can be found at \link{Comparison}.
}
\examples{
x <- biginteger(5)
y <- bigfloat(2)

x < y
x > y
x <= y
x >= y
x == y
x != y
}
\seealso{
Other bignum operations: 
\code{\link{bignum-arith}},
\code{\link{bignum-math}},
\code{\link{bignum-special}}
}
\concept{bignum operations}
