% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{bignum-constants}
\alias{bignum-constants}
\alias{NA_biginteger_}
\alias{NA_bigfloat_}
\alias{bigpi}
\title{Constants}
\usage{
NA_biginteger_

NA_bigfloat_

bigpi
}
\value{
A \code{\link{biginteger}} or \code{\link{bigfloat}} vector of length 1.
}
\description{
\code{NA_biginteger_} and \code{NA_bigfloat_} support missing values.

\code{bigpi} is a higher precision version of \code{\link{pi}}.
}
\examples{
NA_biginteger_

NA_bigfloat_

bigpi
}
\seealso{
\code{\link{NA}} and \code{\link{pi}} are the base constants.
}
\keyword{datasets}
