% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{bignum-format}
\alias{bignum-format}
\alias{format.bignum_biginteger}
\alias{format.bignum_bigfloat}
\title{Formatting}
\usage{
\method{format}{bignum_biginteger}(x, ...)

\method{format}{bignum_bigfloat}(x, scientific = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{biginteger}} or \code{\link{bigfloat}} vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{scientific}{Logical specifying whether to use scientific format.}
}
\description{
Formatting
}
