% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bignum-package.R
\docType{package}
\name{bignum-package}
\alias{bignum}
\alias{bignum-package}
\title{bignum: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Hall \email{david.hall.physics@gmail.com} (\href{https://orcid.org/0000-0002-2193-0480}{ORCID})

}
\keyword{internal}
