% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigfloat.R, R/biginteger.R, R/bignum-package.R
\name{vec_arith.bignum_bigfloat}
\alias{vec_arith.bignum_bigfloat}
\alias{vec_arith.bignum_biginteger}
\alias{bignum-vctrs}
\title{Internal vctrs methods}
\usage{
\method{vec_arith}{bignum_bigfloat}(op, x, y, ...)

\method{vec_arith}{bignum_biginteger}(op, x, y, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
