% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigfloat.R
\name{bigfloat}
\alias{bigfloat}
\alias{as_bigfloat}
\alias{is_bigfloat}
\title{High-Precision Numeric Vectors}
\usage{
bigfloat(...)

as_bigfloat(x)

is_bigfloat(x)
}
\arguments{
\item{...}{Numeric or character vectors.}

\item{x}{Object to be coerced or tested.}
}
\description{
Creates or tests for high-precision floating-point numeric vectors.
}
\examples{
bigfloat(1 / 3)
bigfloat(1) / 3
}
\seealso{
\code{\link{NA_bigfloat_}} to represent missing values.

\code{\link[=format.bignum_bigfloat]{format()}} for pretty printing.

\code{vignette("operations")} for supported operations.
}
