% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biginteger.R
\docType{data}
\name{biginteger}
\alias{biginteger}
\alias{as_biginteger}
\alias{is_biginteger}
\alias{format.bignum_biginteger}
\alias{NA_biginteger_}
\title{Arbitrary-Precision Integer Vectors}
\usage{
biginteger(x = character())

as_biginteger(x, ...)

is_biginteger(x)

\method{format}{bignum_biginteger}(x, ...)

NA_biginteger_
}
\arguments{
\item{x}{Object to be coerced or tested}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Creates or tests for arbitrary-precision integer vectors.

The \code{NA_biginteger_} constant supports missing values.
}
\seealso{
\code{\link[=bigfloat]{bigfloat()}}, \code{\link[=integer]{integer()}}
}
\keyword{datasets}
