% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biginteger.R
\name{biginteger}
\alias{biginteger}
\alias{as_biginteger}
\alias{is_biginteger}
\title{Arbitrary-Precision Integer Vectors}
\usage{
biginteger(...)

as_biginteger(x)

is_biginteger(x)
}
\arguments{
\item{...}{Numeric or character vectors.}

\item{x}{Object to be coerced or tested.}
}
\description{
Creates or tests for arbitrary-precision integer vectors.
}
\seealso{
\code{\link{NA_biginteger_}} to represent missing values.

\code{\link[=format.bignum_biginteger]{format()}} for pretty printing.

Operations supported by biginteger vectors: \link{bignum-compare}, \link{bignum-arith},
\link{bignum-math}.
}
