% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-arith.R
\name{bignum-arith}
\alias{bignum-arith}
\title{Arithmetic operations}
\description{
\code{\link[=biginteger]{biginteger()}} and \code{\link[=bigfloat]{bigfloat()}} vectors support infix operators:
\code{+}, \code{-}, \code{*}, \code{/}, \code{^}, \code{\%\%} and \code{\%/\%}.
}
