% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-arith.R
\name{bignum-arith}
\alias{bignum-arith}
\title{Arithmetic Operations}
\value{
These arithmetic operations are \strong{type-stable}, which means the output type
depends only on the input types (not the input values). A biginteger vector
is returned when the result must be an integer (e.g., addition of two integers).
Otherwise a bigfloat vector is returned.

The following table summarizes the return type for each combination, where
"integer-like" refers to integer and biginteger vectors and "float-like"
refers to double and bigfloat vectors.\tabular{lclcl}{
   \strong{Input 1} \tab \strong{Operator} \tab \strong{Input 2} \tab  \tab \strong{Result} \cr
   Integer-like \tab +, -, *, ^, \%\%, \%/\% \tab Integer-like \tab -> \tab biginteger \cr
   Integer-like \tab / \tab Integer-like \tab -> \tab bigfloat \cr
   Integer-like \tab +, -, *, /, ^, \%\% \tab Float-like \tab -> \tab bigfloat \cr
   Integer-like \tab \%/\% \tab Float-like \tab -> \tab biginteger \cr
   Float-like \tab +, -, *, /, ^, \%\% \tab Integer-like \tab -> \tab bigfloat \cr
   Float-like \tab \%/\% \tab Integer-like \tab -> \tab biginteger \cr
   Float-like \tab +, -, *, /, ^, \%\% \tab Float-like \tab -> \tab bigfloat \cr
   Float-like \tab \%/\% \tab Float-like \tab -> \tab biginteger \cr
}
}
\description{
\code{\link{biginteger}} and \code{\link{bigfloat}} vectors support the standard
\link[=Arithmetic]{arithmetic operations}.\preformatted{+ x
- x
x + y
x - y
x * y
x / y
x ^ y
x \%\% y
x \%/\% y
}
}
\section{Recycling rules}{

The two inputs have \strong{compatible sizes} if:
\itemize{
\item Both input vectors have the same length, or
\item One input is a scalar (length 1) which is recycled.
}
}

