% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-compare.R
\name{bignum-compare}
\alias{bignum-compare}
\title{Comparison Operations}
\value{
A logical vector.
}
\description{
\code{\link{biginteger}} and \code{\link{bigfloat}} vectors support the standard
\link[=Comparison]{comparison operations}.\preformatted{x < y
x > y
x <= y
x >= y
x == y
x != y
}
}
\section{Recycling rules}{

The two inputs have \strong{compatible sizes} if:
\itemize{
\item Both input vectors have the same length, or
\item One input is a scalar (length 1) which is recycled.
}
}

\seealso{
Base R \link[=Comparison]{comparison operations}.
}
