% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{bignum-format}
\alias{bignum-format}
\alias{format.bignum_biginteger}
\alias{format.bignum_bigfloat}
\title{Formatting}
\usage{
\method{format}{bignum_biginteger}(x, ...)

\method{format}{bignum_bigfloat}(x, ...)
}
\arguments{
\item{x}{A \code{\link{biginteger}} or \code{\link{bigfloat}} vector.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Character vector
}
\description{
Format a bignum vector for pretty printing.
}
\examples{
x <- biginteger(2)^100L

# default shows maximum precision
format(x)
}
