% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigNum-package.R
\docType{package}
\name{BigNum-package}
\alias{BigNum}
\alias{BigNum-package}
\title{BigNum: An S7 Showcase}
\description{
A toy package designed to showcase and serve as a reference/tutorial for S7 development. To be used in conjunction with the S7 package website.
}
\author{
\strong{Maintainer}: Visruth Srimath Kandali \email{public@visruth.com} (\href{https://orcid.org/0009-0005-9097-0688}{ORCID})

}
\keyword{internal}
