% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bignum-package.R, R/vctrs-arith.R
\name{bignum-vctrs}
\alias{bignum-vctrs}
\alias{vec_arith.bignum_vctr}
\alias{vec_arith.bignum_biginteger}
\title{Internal vctrs methods}
\usage{
\method{vec_arith}{bignum_vctr}(op, x, y, ...)

\method{vec_arith}{bignum_biginteger}(op, x, y, ...)
}
\value{
\code{vec_arith.bignum_vctr()} returns a \code{\link{bigfloat}} vector.

\code{vec_arith.bignum_biginteger()} returns a \code{\link{biginteger}} vector.
}
\description{
Internal vctrs methods
}
\keyword{internal}
