% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{seq.bignum_vctr}
\alias{seq.bignum_vctr}
\title{Sequences: bignum}
\usage{
\method{seq}{bignum_vctr}(from, to = NULL, by = NULL, length.out = NULL, along.with = NULL, ...)
}
\description{
This is a bignum method for the \code{\link[=seq]{seq()}} generic.

Using \code{seq()} on bignum objects always retains the type of \code{from}.

When calling \code{seq()}, exactly two of the following must be specified:
\itemize{
\item \code{to}
\item \code{by}
\item Either \code{length.out} or \code{along.with}
}
}
